Crate cu29_log

Re-exports

pub use cu29_value as value;

Structs

CuLogEntry

This is the basic structure for a log entry in Copper.

Constants

ANONYMOUS

MAX_LOG_PARAMS_ON_STACK

Functions

default_log_index_dir

Convenience function to returns the default path for the log index directory.

format_logline

Text log line formatter.

rebuild_logline

Rebuild a log line from the interned strings and the CuLogEntry. This basically translates the world of copper logs to text logs.