Crate cu29_log

Macros

__cu29_defmt_debug

__cu29_defmt_error

__cu29_defmt_info

__cu29_defmt_warn

defmt_debug

defmt_error

defmt_info

defmt_warn

Structs

CuLogEntry

This is the basic structure for a log entry in Copper.

Enums

CuLogLevel

Log levels for Copper.

Constants

ANONYMOUS

MAX_LOG_PARAMS_ON_STACK

Functions

format_logline

Text log line formatter. Only available on std. TODO(gbin): Maybe reconsider that at some point

rebuild_logline

Rebuild a log line from the interned strings and the CuLogEntry. This basically translates the world of copper logs to text logs.